Estimator#shipping_rates sets no rates on packages
As of solidusio/solidus@0684205c22d91835f4ba0e87efe6d0fb3b20c5fa
Spree::Stock:Package does not maintain a list of its own shipping rates.

The Estimator is not required to associate the shipping rates to the
shippment. Meerly collect eligible rates and return them to the
Stock::Coordinator is responsible for associating the rates to the
actual shipment [here](0684205c22d91835f4ba0e87efe6d0fb3b20c5fa).

This is a slight change of behaviour to stay compatible with future
versions of solidus, but it does not break compatibility with older
versions of solidus.