Update Client Instantiation for EasyPost 5.0
The EasyPost 5.0 gem changed the way the client object is instantiated,
requiring the creation of a client that holds the API key and routing
all easypost calls through that client.

This change updates the gem's config to make the API key part of the
configuration, and create a client on instantiation that can be used to
make calls to Easypost.