Update EasyPost Postage Label
Use SolidusEasypost Client to purchase a Shipment in order to populate the tracking_code and postage_label attributes for a specific shipment. The static postage_label_url in the test is adjusted to match the s3 url for the uploaded postage_label png file.
The uploaded easypost_postage_label_url includes the datestamp on which the file is uploaded and filename.
This adjusted test ensures the dynamic behaviour of easypost_postage_label_url generation.